latest updates... a lot

+++ Fallen-8/Algorithms/Path/IShortestPathAlgorithm.cs
@@ -4,7 +4,7 @@
 // Author:
 //       Henning Rauch <Henning@RauchEntwicklung.biz>
 // 
-// Copyright (c) 2012 Henning Rauch
+// Copyright (c) 2012-2015 Henning Rauch
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -54,8 +54,8 @@ namespace NoSQL.GraphDB.Algorithms.Path
         /// <param name="vertexCost">The vertex cost delegate.</param>
         /// <returns>Paths</returns>
         List<Path> Calculate(
-            Int64 sourceVertexId, 
-            Int64 destinationVertexId,
+            Int32 sourceVertexId, 
+            Int32 destinationVertexId,
             Int32 maxDepth = 1,
             Double maxPathWeight = Double.MaxValue,
             Int32 maxResults = 1,
